Widen lost_words on the overwrite-retry path
The overwrite-race fix adds a third lost_events call site. It was written
against the int signature and needs the same uint64_t as the other two.

+++ otherlibs/runtime_events/runtime_events_consumer.c
@@ -514,7 +514,8 @@ caml_runtime_events_read_poll(struct caml_runtime_events_cursor *cursor,
           atomic_load_acquire(&runtime_events_buffer_header->ring_head);
 
         if (ring_head > cursor->current_positions[domain_num]) {
-          int lost_words = ring_head - cursor->current_positions[domain_num];
+          uint64_t lost_words =
+            ring_head - cursor->current_positions[domain_num];
           cursor->current_positions[domain_num] = ring_head;
 
           if (cursor->lost_events) {
